fix(sonar): clear 5 code-quality issues + annotate 1 Play control-flow FP
Triage of the open SonarQube issues for abundent:jclaw (gate was ERROR on
new_reliability_rating=C). All sites are in the recent video/audio/vLLM code.

Fixes:
- java:S1192  ApiProvidersController — extract BASE_URL_SUFFIX (".baseUrl" x4)
- java:S1192  AgentRunner — extract OUTCOME_ERROR ("error" outcome label x3)
- java:S135   ApiProvidersController.videoModels — merge two loop guards into one
- java:S6126  AgentRunnerStreamingPathTest — SSE mock concat to a text block
- java:S1905  VideoUrlAdapter.maxInlineBytes — drop redundant (long) cast

False positive:
- java:S2259  ApiProvidersController.reachable — the null guard halts via Play's
  renderJSON() throwing, which Sonar doesn't model. Added @SuppressWarnings to
  match the sibling discoverModels; marked FALSE_POSITIVE in Sonar.

Backend suite green (284 classes); frontend green (776 tests). Gate now OK.

+++ app/services/video/VideoUrlAdapter.java
@@ -34,7 +34,7 @@ public final class VideoUrlAdapter {
     public static long maxInlineBytes() {
         int mb = Math.clamp(ConfigService.getInt(CFG_MAX_INLINE_MB, DEFAULT_MAX_INLINE_MB),
                 MAX_INLINE_MB_MIN, MAX_INLINE_MB_MAX);
-        return (long) mb * 1024L * 1024L;
+        return mb * 1024L * 1024L;
     }
 
     /** True when the clip is small enough to inline as a base64 {@code video_url}. */
